DRY up blog post code
